==> DrugProfiles_2/1_4_Filtering_Fitting.ipynb <==
## Filtering

import pandas as pd
import numpy as np
import os
import gc
import matplotlib.pyplot as plt
import time

from functions.filtering import *
from functions.plotting import *
from functions.fitting import *

_FOLDER = "database/"
_FOLDER_2 = "figures/"
_FOLDER_3 = "results/"
SAVE_FIGURES = False

R2_limit = 0.99

### Original data

drug_curves = pd.read_csv(_FOLDER+"normalised_dose_response_data.csv")

if "Unnamed: 0" in drug_curves:
    drug_curves.drop("Unnamed: 0", axis=1, inplace =True)

col_to_drop = ["per_slope_change_"+str(i) for i in range(8)]+\
            ["slope_" + str(i) for i in range(9)]    
drug_curves.drop(col_to_drop, axis=1, inplace=True)    
    
conc_columns= ["fd_num_"+str(i) for i in range(10)]
response_norm = ['norm_cells_'+str(i) for i in range(10)]
CCL_names = dict(zip(drug_curves["COSMIC_ID"], drug_curves["CELL_LINE_NAME"]))
df= pd.read_csv(_FOLDER+'Drug_Features.csv')
drug_names = dict(zip(df["Drug ID"].values, df["Drug Name"].values))
del df
drug_curves["drug_name"] = drug_curves["DRUG_ID"].map(drug_names)
drug_curves["CCL_name"] = drug_curves["COSMIC_ID"].map(CCL_names)
drug_curves.shape

## Filtering 1: 4 stage filtering

# Description of filtering_sigmoid_curves:
"""
    filtering_scenario = [1,2,3,4]
    1. Ensure that all the response are less than 1

    2. Ensure that first and last points form plateus
    the minimal number of points are specified in the function arguments
    by default, two points for both lpateus are considered
    tolerance =0.05 values to ensure the points form a plateu
    first_columns_to_compare = [1, 2]  - first two columns for plateu
    last_columns_to_compare = [-1, -2] - last two columns for plateu

    3. Specify location of the plateus - first_points_lower_limit and last_points_upper_limit

    4. Cutting off ambiqueos data:
    Among all "middle" datapoints a subsequent point should not be higher than antecedent by 0.2
"""

%%time
# difference between middle points 0.2
df_filt_1234 = filtering_sigmoid_curves(drug_curves, filtering_scenario=[1,2,3,4], \
                        response_columns = response_norm, \
                        first_points_lower_limit = 0.8, last_points_upper_limit = 0.2, 
                         middle_points_limit = -0.2)
df_filt_1234.to_csv(_FOLDER_3+"filt_1234_02.csv", index=False)

%%time
df = df_filt_1234.copy()
fitting_function = "sigmoid_4_param"

r2, fit_param = fitting_column(df, df.index, x_columns=conc_columns, y_columns= response_norm,
                               fitting_function = fitting_function, default_param=True)
df[fitting_function+"_r2"] = r2
df[fitting_function] = fit_param
df = df[df[fitting_function+"_r2"]>R2_limit]
print(df_filt_1234.shape, df.shape)
df.to_csv(_FOLDER_3+"fit_1234_02.csv", index=False)

## Filtering 2:  auc>0.7 and spearman_r<0

# description of auc_filtration
"""
    1. Remove all the curves where the normalised response value is greater than one at zero dosage.
    2. Leave only those curves with an Area Under the Curve (AUC) >0.7.  
    3. Compute the Spearman correlation coefficient between the normalised response 
    and the scaled dosage (so the x-axis and the y-axis). 
    4. Further remove the curves for which the Spearman correlation coefficient is zero or positive. 
    5. Cut off samples with last response above 0.2
"""

%%time
df_filt_auc = pd.read_csv(_FOLDER_3+"filt_auc_02.csv")
df = df_filt_auc.copy()
fitting_function = "sigmoid_4_param"

r2, fit_param = fitting_column(df, df.index, x_columns=conc_columns, y_columns= response_norm,
                               fitting_function = fitting_function, default_param=True)
df[fitting_function+"_r2"] = r2
df[fitting_function] = fit_param
df = df[df[fitting_function+"_r2"]>R2_limit]
print(df_filt_auc.shape, df.shape)
df.to_csv(_FOLDER_3+"fit_auc_02.csv", index=False)

## Filtering 3: direct fitting

functions = [
             "fsigmoid", 
             "sigmoid_2_param",
             "sigmoid_3_param",
             "sigmoid_4_param",
             "logistic_4_param",
            "ll4_4_param",
            "ll4R_4_param",
            "logLogist_3_param"]

%%time
df_no_filt = compare_fitting_functions(drug_curves, functions, conc_columns, response_norm, 
                                       save_file_name = _FOLDER_3 +"fit_no_filt.csv")
