==> DrugProfiles_2/3_0_Merging_all_data.ipynb <==
## Merging filtered and fit data with CCL properties

import pandas as pd
import numpy as np

from scipy.optimize import curve_fit
import scipy.optimize as opt
from sklearn.metrics import r2_score

from tqdm import tqdm
import warnings
warnings.filterwarnings("ignore")

import os
_FOLDER = "database/"
_FOLDER_2 = "results/"

from functions.merging import *

cell_features = pd.read_csv(_FOLDER +"Cell_Line_Features_PANCAN_simple_MOBEM.tsv", sep="\t")
drug_features = pd.read_csv(_FOLDER_2+ "drug_features_with_pubchem_properties.csv")

### Merging with filtered and fitted data

df_list = ["fit_1234_02", "fit_auc_02", "filt_1234_02","filt_auc_02"]
df_list_2 = [_FOLDER_2 + x + ".csv" for x in df_list]

df = pd.read_csv(df_list_2[0])
df.head(2)

i=0
df = pd.read_csv(df_list_2[i])
merged_df = merge_drug_cells(df, cell_features, drug_features,save_CCL_properties=True, _FOLDER_to_save=_FOLDER_2)

for i in range(len(df_list_2)):
    df = pd.read_csv(df_list_2[i])
    merged_df = merge_drug_cells(df, cell_features, drug_features)
    new_file_name = _FOLDER_2 + df_list[i]+"_merged.csv"
    merged_df.to_csv(new_file_name, index=False)
    print(new_file_name)
    print(f"\t Drug Profiles+CCL {df.shape}, Merged Data: {merged_df.shape}")
    gr = merged_df.groupby("DRUG_ID").agg(count=("CELL_LINE_NAME", "count"))
    print("\tN unique drugs:", gr.shape[0])
    print("\tN of 10+ drug records:", gr[gr["count"]>10].shape[0])
