==> DrugProfiles_2/4_Train_Validation_Test_data.ipynb <==
## Merging filtered and fit data with CCL properties

import pandas as pd
import numpy as np

from tqdm import tqdm
import warnings
warnings.filterwarnings("ignore")

import os
_FOLDER = "results/"

df_list = ["fit_1234_02_merged.csv", "fit_auc_02_merged.csv", "fit_no_filt_merged.csv"]

df = pd.read_csv(_FOLDER + "fit_1234_02_merged.csv")
df.head(2)

df.shape[0], df["DRUGID_COSMICID"].nunique()

### Split into train and test data with more than 10 record per drug

gr = df.groupby("DRUG_ID").size()
drugs = gr[gr>10].index
print("Number of drugs with more than 10 profiles:", len(drugs))        

train_ratio = 0.8
train = pd.DataFrame()
test = pd.DataFrame()
np.random.seed(123)
indexes = np.random.permutation(df.index)

for drug_id in drugs:
    df_i = df[df["DRUG_ID"]==drug_id]
    indexes = np.random.permutation(df_i.index)
    train_size = int(df_i.shape[0]*train_ratio)
    indexes_train = indexes[:train_size]
    indexes_test = indexes[train_size:]
    train = pd.concat([train, df_i.loc[indexes_train, :]])
    test = pd.concat([test, df_i.loc[indexes_test, :]])
    
train.set_index("DRUGID_COSMICID", inplace=True)
test.set_index("DRUGID_COSMICID", inplace=True)

print(train.shape, test.shape)

with open(_FOLDER +"train_10.txt", 'w') as f:
    for s in train.index:
        f.write(str(s) + '\n')
        
with open(_FOLDER +"test_10.txt", 'w') as f:
    for s in test.index:
        f.write(str(s) + '\n')

### Possible scenarios:
- drug-by-drug learning, i.e separate models for each fitting coefficients for each drug
- all drugs - one model for each coefficients for all drugs
- all data - one model for each response for all drugs and drug profiles
