==> DrugProfiles_2/8_1_Tuned_Hyperparameters.ipynb <==
## Processing hyperparameters from text format

import pandas as pd
import numpy as np
_FOLDER_2 = "results/"

# Coefficient 1
coef=1
best_parameters=  ["DRUG_ID	best_alpha3	best_gamma3	best_coef03	best_degree3",
"170	100.0	0.10000	5.0	1.0",
"173	100.0	0.01000	5.0	3.0",
"180	100.0	0.00001	0.5	1.0",
"200	100.0	0.00001	5.0	1.0",
"219	10.0	0.00001	0.5	2.0",
"272	500.0	0.00010	5.0	2.0",
"273	0.5	0.01000	1.0	1.0",
"274	1.0	0.00001	0.5	5.0",
"276	1.0	0.00001	5.0	5.0",
"328	5.0	0.00010	0.5	3.0",
"346	1.0	0.01000	0.5	1.0"]

param_dict = {}
parameters_names = best_parameters[0].split("\t")
for i in range(1, len(best_parameters)):
    for j,param in enumerate(parameters_names):
        if param in param_dict:
            param_dict[param].append(best_parameters[i].split("\t")[j])
        else:
            param_dict[param] = [best_parameters[i].split("\t")[j]]
            
hyperparameters = pd.DataFrame(param_dict)
hyperparameters["DRUG_ID"] = np.int32(hyperparameters["DRUG_ID"])
hyperparameters.set_index("DRUG_ID", inplace= True)
for col in hyperparameters.columns:
    hyperparameters[col] = np.float32(hyperparameters[col])

hyperparameters.to_csv(_FOLDER_2+"hyperparameters_coef"+str(coef)+"_alg1.csv")    
hyperparameters

# Coefficient 2
coef=2
best_parameters=  ["DRUG_ID	best_C3	best_epsilon3	best_coef03",
"170	0.1	0.010	0.01",
"173	0.1	0.001	0.01",
"180	0.1	0.010	0.01",
"200	0.1	0.010	0.01",
"219	0.5	0.010	0.01",
"272	0.1	0.001	0.01",
"273	0.1	0.001	0.01",
"274	0.1	0.010	0.01",
"276	0.1	0.010	0.01",
"328	0.1	0.001	0.01",
"346	0.1	0.001	0.01"]

param_dict = {}
parameters_names = best_parameters[0].split("\t")
for i in range(1, len(best_parameters)):
    for j,param in enumerate(parameters_names):
        if param in param_dict:
            param_dict[param].append(best_parameters[i].split("\t")[j])
        else:
            param_dict[param] = [best_parameters[i].split("\t")[j]]
            
hyperparameters = pd.DataFrame(param_dict)
hyperparameters["DRUG_ID"] = np.int32(hyperparameters["DRUG_ID"])
hyperparameters.set_index("DRUG_ID", inplace= True)
for col in hyperparameters.columns:
    hyperparameters[col] = np.float32(hyperparameters[col])

hyperparameters.to_csv(_FOLDER_2+"hyperparameters_coef"+str(coef)+"_alg1.csv")    
hyperparameters

coef=2
name =_FOLDER_2+"hyperparameters_coef"+str(coef)+"_alg1.csv"
name
pd.read_csv(_FOLDER_2+"hyperparameters_coef2_alg1.csv")

pd.read_csv(_FOLDER_2+"hyperparameters_coef2_alg1.csv")

# Coefficient 3
coef=3
best_parameters=  ["DRUG_ID	best_C3	best_epsilon3",
"170	0.10	5.000",
"173	0.10	0.001",
"180	5.00	5.000",
"200	0.01	0.100",
"219	0.10	1.000",
"272	0.01	2.000",
"273	0.01	2.000",
"274	0.10	0.100",
"276	0.01	1.000",
"328	0.01	2.000",
"346	0.01	0.001"]
param_dict = {}
parameters_names = best_parameters[0].split("\t")
for i in range(1, len(best_parameters)):
    for j,param in enumerate(parameters_names):
        if param in param_dict:
            param_dict[param].append(best_parameters[i].split("\t")[j])
        else:
            param_dict[param] = [best_parameters[i].split("\t")[j]]
            
hyperparameters = pd.DataFrame(param_dict)
hyperparameters["DRUG_ID"] = np.int32(hyperparameters["DRUG_ID"])
hyperparameters.set_index("DRUG_ID", inplace= True)
for col in hyperparameters.columns:
    hyperparameters[col] = np.float32(hyperparameters[col])

hyperparameters.to_csv(_FOLDER_2+"hyperparameters_coef"+str(coef)+"_alg1.csv")    
hyperparameters

# Coefficient 4
coef=4
best_parameters=  ["DRUG_ID	best_alpha3	best_gamma3	best_coef03",
"170	50.0	0.00001	0.01",
"173	0.1	1.00000	0.01",
"180	10.0	0.00010	0.01",
"200	0.1	0.00010	0.01",
"219	5.0	0.00001	0.01",
"272	10.0	0.00001	0.01",
"273	50.0	0.01000	0.01",
"274	100.0	0.00001	0.01",
"276	50.0	0.00001	0.01",
"328	0.1	0.00010	0.01",
"346	10.0	0.00001	0.01"]
param_dict = {}
parameters_names = best_parameters[0].split("\t")
for i in range(1, len(best_parameters)):
    for j,param in enumerate(parameters_names):
        if param in param_dict:
            param_dict[param].append(best_parameters[i].split("\t")[j])
        else:
            param_dict[param] = [best_parameters[i].split("\t")[j]]
            
hyperparameters = pd.DataFrame(param_dict)
hyperparameters["DRUG_ID"] = np.int32(hyperparameters["DRUG_ID"])
hyperparameters.set_index("DRUG_ID", inplace= True)
for col in hyperparameters.columns:
    hyperparameters[col] = np.float32(hyperparameters[col])

hyperparameters.to_csv(_FOLDER_2+"hyperparameters_coef"+str(coef)+"_alg1.csv")    
hyperparameters
